# client: add WhereAny() and let Where() take multiple models