--- a/client/client.go
+++ b/client/client.go
@@ -1312,9 +1312,14 @@ func (o *ovsdbClient) List(ctx context.Context, result interface{}) error {
 	return primaryDB.api.List(ctx, result)
 }
 
-//Where implements the API interface's Where function
-func (o *ovsdbClient) Where(m model.Model, conditions ...model.Condition) ConditionalAPI {
-	return o.primaryDB().api.Where(m, conditions...)
+// Where implements the API interface's Where function
+func (o *ovsdbClient) Where(m model.Model) ConditionalAPI {
+	return o.primaryDB().api.Where(m)
+}
+
+// WhereAny implements the API interface's WhereAny function
+func (o *ovsdbClient) WhereAny(m model.Model, conditions ...model.Condition) ConditionalAPI {
+	return o.primaryDB().api.WhereAny(m, conditions...)
 }
 
 //WhereAll implements the API interface's WhereAll function