There are cases where we'd like to take a list of models and return a list of results without grabbing the row cache lock for every single item. Batch them up into a Where() condition that can then be List()-ed right after, which is about 2x faster than Get()-ing each item individually.

Before:
```
BenchmarkAPIListMultiple/multiple_results_one_at_a_time_with_Get-8         	     212	   5534355 ns/op
BenchmarkAPIListMultiple/multiple_results_one_at_a_time_with_Get-8         	     212	   5640041 ns/op
BenchmarkAPIListMultiple/multiple_results_one_at_a_time_with_Get-8         	     206	   5680483 ns/op
```
After:
```
BenchmarkAPIListMultiple/multiple_results_in_a_batch_with_WhereAny-8       	     420	   2795212 ns/op
BenchmarkAPIListMultiple/multiple_results_in_a_batch_with_WhereAny-8       	     412	   2895229 ns/op
BenchmarkAPIListMultiple/multiple_results_in_a_batch_with_WhereAny-8       	     412	   2864806 ns/op
```

This PR also adds a WhereAny() call that acts like WhereAll() except with OR semantics, which is equivalent to the previous Where() call including conditions. Thus we now have:

1. Where()  - accepts one or more models and matches on indexes and model data
2. WhereAny() - takes a model and one or more conditions, matching each condition with logical OR
3. WhereAll()  - takes a model and one or more conditions, matching all conditions with a logical AND


@jcaamano @flavio-fernandes @trozet @tssurya