@dcbw 2022-06-22
> I have a feeling that this `WhereMany` is how `Where` should have been all along. What we could do is have `WhereAny` instead of `WhereMany`, that basically deprecates `Where`.

@jcaamano updated to be WhereAny(). I still need to add testcases.