@jcaamano 2022-08-11
Wan't to run it by ovn-k to discard issues?

> @dcbw could you add tests where you pass models of different types, in the same places you added new tests? That should fail in some way, right?

Just cross-checking you read this. What I see is that you use the first model of the list to know which type it is and get the RowCache matching that type. And then in `RowsByModels` you skip any model that does not match that type. I am considering if an error would be more appropriate, as skipping can lead to the interpretation that the request made sense but the looked after models did not exist.