@dcbw 2022-08-22
> And then in `RowsByModels` you skip any model that does not match that type.

@jcaamano RowsByModels() doesn't return an error; I could make it return one though.