@dcbw 2022-08-22
@jcaamano how about now? changed rowsByModel() to return an error and followed that through.

Also added testcases as requested to ensure model lists return an error when the models aren't the same type as the first element.