@dcbw 2022-08-23
@jcaamano ran a 120 density-light test (which is what exercises this most) downstream and got

a7da83bf-node-density-20220823 (P99=4.19s AVG=3.04s)

which is slightly better than usual. Timings show that:

```
I0823 20:08:22.492399       1 router.go:897] ####### CreateOrUpdateNATsOps; (getRouterNats 234 took router: 134.883µs, nats: 507.978µs); reconcile took 19.239µs
I0823 20:08:22.532599       1 router.go:897] ####### CreateOrUpdateNATsOps; (getRouterNats 237 took router: 129.662µs, nats: 555.431µs); reconcile took 21.32µs
I0823 20:08:22.587449       1 router.go:897] ####### CreateOrUpdateNATsOps; (getRouterNats 239 took router: 129.239µs, nats: 562.158µs); reconcile took 20.164µs
I0823 20:08:22.637912       1 router.go:897] ####### CreateOrUpdateNATsOps; (getRouterNats 236 took router: 165.132µs, nats: 484.924µs); reconcile took 18.919µs
I0823 20:08:22.691104       1 router.go:897] ####### CreateOrUpdateNATsOps; (getRouterNats 235 took router: 126.207µs, nats: 595.055µs); reconcile took 39.61µs
I0823 20:08:22.739018       1 router.go:897] ####### CreateOrUpdateNATsOps; (getRouterNats 237 took router: 129.993µs, nats: 558.256µs); reconcile took 19.946µs
```

looking up 230+ NAT objects in a batch almost never take more than 1ms while before they could take a *lot* more time than that.